Add dialog boxes for Boyd
